_delphi_utils_python/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "delphi-utils"
version = "0.3.25"
description = "Shared Utility Functions for Indicators"
readme = "README.md"
requires-python = "== 3.8.*"
license = { text = "MIT License" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "boto3",
    "covidcast",
    "cvxpy",
    "epiweeks",
    "gitpython",
    "importlib_resources>=1.3",
    "numpy",
    "pyarrow",
    "pandas>=1.1.0",
    "requests",
    "slackclient",
    "scs<3.2.6",                # TODO: remove this ; it is a cvxpy dependency, and the excluded version appears to break our jenkins build. see: https://github.com/cvxgrp/scs/issues/283
    "structlog",
    "xlrd",                     # needed by Pandas to read Excel files
]

[project.urls]
Homepage = "https://github.com/cmu-delphi/covidcast-indicators"

[project.optional-dependencies]
dev = [
    "darker[isort]~=2.1.1",
    "pylint==2.8.3",
    "pytest",
    "pydocstyle",
    "pytest-cov",
    "mock",
    "moto~=4.2.14",
    "requests-mock",
    "freezegun",
]
flash = ["scipy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["delphi_utils"]
namespaces = true

[tool.setuptools.package-data]
"delphi_utils.data" = ["20*/*.csv"]
